revert a few minor fixes to not include in this prod release. Will th…

+++ cpu-app/Models/program-application/DynamicsProgramApplicationProgramPost.cs
@@ -10,7 +10,11 @@ namespace Gov.Cscp.Victims.Public.Models
         public string fortunecookietype { get { return "Microsoft.Dynamics.CRM.vsd_program"; } }
         public bool vsd_costshare { get; set; }
         public bool vsd_cpu_programstaffsubcontracted { get; set; }
-        public float? vsd_cpu_numberofhours { get; set; }
+
+        //This isn't an int, but not including this update in this prod release, will let the fix get tested for next release
+        //it should just be removed since it is a read only field
+        public int? vsd_cpu_numberofhours { get; set; }
+        // public float? vsd_cpu_numberofhours { get; set; }
         public int? vsd_cpu_per { get; set; }
         public int? vsd_totaloncallstandbyhours { get; set; }
         public int? vsd_totalscheduledhours { get; set; }
